// src/api/generated/models/productPublicListResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * NEST Boilerplate Api Documentation
 * Boilerplate of NEST project Api Documentation to test and review APIs
 * OpenAPI spec version: 1.0
 */
import type { PaginationMetaResponse } from "./paginationMetaResponse";
import type { ProductPublicResponse } from "./productPublicResponse";

export interface ProductPublicListResponse {
  data: ProductPublicResponse[];
  meta: PaginationMetaResponse;
}