// src/api/generated/models/productsControllerGetApprovedProductsV1Params.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * NEST Boilerplate Api Documentation
 * Boilerplate of NEST project Api Documentation to test and review APIs
 * OpenAPI spec version: 1.0
 */
import type { ProductsControllerGetApprovedProductsV1SortOrder } from "./productsControllerGetApprovedProductsV1SortOrder";
import type { ProductsControllerGetApprovedProductsV1Status } from "./productsControllerGetApprovedProductsV1Status";

export type ProductsControllerGetApprovedProductsV1Params = {
  isPagination?: boolean;
  page?: number;
  limit?: number;
  search?: string;
  sortOrder?: ProductsControllerGetApprovedProductsV1SortOrder;
  sortBy?: string;
  status?: ProductsControllerGetApprovedProductsV1Status;
  vendorId?: string;
  name?: string;
  isDeleted?: boolean;
};